fix(client): Fix theme barrel exports
