--- apps/client/src/app/themes/Themes.ts
@@ -1,25 +1,0 @@
-import { createTheme } from '@mui/material/styles';
-
-export const DarkTheme = createTheme({
-    palette: {
-        mode: 'dark',
-        primary: {
-            main: '#5ee25c',
-        },
-        secondary: {
-            main: '#f50057',
-        },
-    },
-});
-
-export const LightTheme = createTheme({
-    palette: {
-        mode: 'light',
-        primary: {
-            main: '#39BC70',
-        },
-        secondary: {
-            main: '#f50057',
-        },
-    },
-});
