+++ apps/client/src/themes/dark.theme.ts
@@ -1,0 +1,13 @@
+import { createTheme } from '@mui/material';
+
+export const DarkTheme = createTheme({
+    palette: {
+        mode: 'dark',
+        primary: {
+            main: '#5ee25c',
+        },
+        secondary: {
+            main: '#f50057',
+        },
+    },
+});
