+++ apps/client/src/themes/index.ts
@@ -1,0 +1,2 @@
+export * from './dark.theme';
+export * from './light.theme';
