+++ apps/client/src/themes/light.theme.ts
@@ -1,0 +1,13 @@
+import { createTheme } from '@mui/material';
+
+export const LightTheme = createTheme({
+    palette: {
+        mode: 'light',
+        primary: {
+            main: '#39BC70',
+        },
+        secondary: {
+            main: '#f50057',
+        },
+    },
+});
